backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
chromadb
openai
python-dotenv
requests
yfinance
finnhub-python
google-cloud-translate
pandas
numpy

